python/IDE/VSCode/PyEventCore/PyEventCore-v1/requirements.txt
alembic==1.18.1
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
billiard==4.2.4
black==26.1.0
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.4
fastapi==0.128.0
flake8==7.3.0
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
kombu==5.6.2
librt==0.7.8
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.4.0
redis==7.1.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.40.0
vine==5.1.0
wcwidth==0.5.0
